runners: lava_runner: Return only the specific test short name
The id() method returns a string identifying the specific test case. This
is usually the full name of the test method, including the module and
class name. For now it is not really useful have this information
printed on this runner, so strip the module and the class name and only
print the specific test name.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>