test_cros_ec_abi: Use complete device name instead of a suffix
To check the sysfs ABI, we use only the suffix ec, fp, pd and tp as
name. We will have in the future, though, other embedded controllers
that will not start with the "cros_" prefix, like the "wilco_ec". To
solve those cases change the code to use the full device name.

Signed-off-by: Fabien Lahoudere <fabien.lahoudere@collabora.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>