test_cros_ec_accel: Check trigger attribute only when is available
Since kernel commit aa984f1ba4a4 ("iio: cros_ec: Register to
cros_ec_sensorhub when EC supports FIFO") the trigger attribute for some
sensors is only exposed when the FIFO feature is not supported. So only
check that attribute when that feature is not supported and on kernels
before that patch was introduced.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>