Add a verbosity flag to add more error context in outputs
The current cros-ec tests as run in LAVA labs are hard to debug when
they fail or when they're skipped because the error cause isn't
propagated to the test output.

This commit adds an additional `--verbosity' flag that can be used to
increase the amount of additional info shown in the test output:

  `--verbosity 1': basic log messages
  `--verbosity 2': basic log messages and full debug info (stacktraces)

By default, no additional error info is printed, ie. if the flag isn't
used we get the same output as we always got.

Signed-off-by: Ricardo Cañuelo <ricardo.canuelo@collabora.com>